Move api request params into separate file & under client folder.
