+++ Src/Notion.Client/Api/Databases/RequestParams/DatabasesListParameters.cs
@@ -1,0 +1,8 @@
+﻿namespace Notion.Client
+{
+    public class DatabasesListParameters : IDatabasesListQueryParmaters
+    {
+        public string StartCursor { get; set; }
+        public string PageSize { get; set; }
+    }
+}
