+++ Src/Notion.Client/Api/Databases/RequestParams/DatabasesQueryParameters.cs
@@ -1,0 +1,12 @@
+﻿using System.Collections.Generic;
+
+namespace Notion.Client
+{
+    public class DatabasesQueryParameters : IDatabaseQueryBodyParameters
+    {
+        public Filter Filter { get; set; }
+        public List<Sort> Sorts { get; set; }
+        public string StartCursor { get; set; }
+        public string PageSize { get; set; }
+    }
+}
