+++ Src/Notion.Client/Api/Databases/RequestParams/IDatabaseQueryBodyParameters.cs
@@ -1,0 +1,10 @@
+﻿using System.Collections.Generic;
+
+namespace Notion.Client
+{
+    public interface IDatabaseQueryBodyParameters : IPaginationParameters
+    {
+        Filter Filter { get; set; }
+        List<Sort> Sorts { get; set; }
+    }
+}
