+++ Src/Notion.Client/Api/Databases/RequestParams/IDatabasesListQueryParmaters.cs
@@ -1,0 +1,6 @@
+﻿namespace Notion.Client
+{
+    public interface IDatabasesListQueryParmaters : IPaginationParameters
+    {
+    }
+}
