+++ Src/Notion.Client/Models/Database/Database.cs
@@ -4,30 +4,6 @@ using System.Collections.Generic;
 
 namespace Notion.Client
 {
-    public interface IDatabasesListQueryParmaters : IPaginationParameters
-    {
-    }
-
-    public class DatabasesListParameters : IDatabasesListQueryParmaters
-    {
-        public string StartCursor { get; set; }
-        public string PageSize { get; set; }
-    }
-
-    public interface IDatabaseQueryBodyParameters : IPaginationParameters
-    {
-        Filter Filter { get; set; }
-        List<Sort> Sorts { get; set; }
-    }
-
-    public class DatabasesQueryParameters : IDatabaseQueryBodyParameters
-    {
-        public Filter Filter { get; set; }
-        public List<Sort> Sorts { get; set; }
-        public string StartCursor { get; set; }
-        public string PageSize { get; set; }
-    }
-
     public class Database
     {
         public string Object => "database";
